backend/requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.1
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.34.144
botocore==1.34.144
cargo==0.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.5.8
colorama==0.4.6
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
databases==0.9.0
datrie==0.8.2
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
dnspython==2.6.1
docker==7.1.0
docopt==0.6.2
easyplotly==0.1.3
email_validator==2.1.1
enforce==0.3.4
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.5
fastjsonschema==2.19.1
fibheap==0.2.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.50.0
frozendict==2.4.0
fsspec==2024.6.1
fuzzingbook==1.2.1
gitdb==4.0.11
GitPython==3.1.42
grammar_graph==0.2.0
grammar_to_regex==0.0.5
graphviz==0.20.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.6
ijson==3.2.3
importlib_metadata==7.0.2
ipython==8.22.2
isla-solver==1.14.4
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lizard==1.17.10
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
networkx==3.2.1
notedown==1.5.1
num2words==0.5.13
numpy==1.26.4
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pandoc-attributes==0.1.7
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.3
pathos==0.3.2
pi==0.1.2
pillow==10.2.0
platformdirs==4.2.0
plotly==5.20.0
pox==0.3.4
ppft==1.7.6.8
prompt-toolkit==3.0.43
proxyorderedset==0.3.5
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
pydantic==2.7.2
pydantic_core==2.18.3
pydot==1.4.2
PyDriller==2.6
Pygments==2.17.2
pygraphviz==1.12
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
reportlab==4.1.0
requests==2.31.0
returns==0.22.0
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.2
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.18.1
shellingham==1.5.4
showast==0.2.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
svglib==1.5.1
tenacity==8.2.3
threadpoolctl==3.3.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
types-pytz==2024.1.0.20240203
types-requests==2.32.0.20240712
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
yapf==0.40.2
z3-solver==4.11.2.0
zipp==3.18.1
